refactor(engine): extract hex_lower to a neutral module
Move hex_lower out of the crypto-adjacent api::signer module into a
private crate-level hex module and revert widening signer to pub(crate).
The rotation error formatter now imports the neutral helper instead of
coupling to signer internals. Addresses CodeRabbit review on #744.
