+++ crates/engine/src/hex.rs
@@ -1,0 +1,26 @@
+//! Lowercase hex encoding — a neutral wire/diagnostic helper.
+//!
+//! Not crypto: a byte-to-hex encoding of already-public bytes (a SEC1 public
+//! key, a compact signature, a scope_id). Lives outside the signing module so
+//! unrelated domains (e.g. rotation diagnostics) can format public bytes
+//! without coupling to `api::signer` internals.
+
+/// Lowercase hex encoding of `bytes`.
+pub(crate) fn hex_lower(bytes: &[u8]) -> String {
+    let mut out = String::with_capacity(bytes.len() * 2);
+    for byte in bytes {
+        out.push(char::from_digit((byte >> 4) as u32, 16).expect("high nibble is < 16"));
+        out.push(char::from_digit((byte & 0x0f) as u32, 16).expect("low nibble is < 16"));
+    }
+    out
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn hex_lower_pads_each_byte() {
+        assert_eq!(hex_lower(&[0x00, 0x0f, 0xa0, 0xff]), "000fa0ff");
+    }
+}
