+++ crates/engine/src/rotation/eager_set.rs
@@ -65,8 +65,8 @@ use std::collections::{BTreeMap, BTreeSet};
 
 use cipherbox_core::seal::ChildScopeRef;
 
-use crate::api::signer::hex_lower;
 use crate::grants::child_index::canonicalize;
+use crate::hex::hex_lower;
 
 /// Why a descendant scope root's own direct-child index could not be
 /// authoritatively obtained. Mirrors the adoption gate's `Rejected` vs `Seam`
